requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==43.0.3
decorator==5.1.1
dj-rest-auth==7.0.1
Django==4.2.16
django-allauth==65.9.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
requests==2.32.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.13
